feat: more JM boilerplate code
fix: coldcard account map import crash

+++ FullyNoded/Wallet Logic/ImportWallet.swift
@@ -25,7 +25,7 @@ class ImportWallet {
         }
         var keypool = Bool()
         var primDescriptor = accountMap["descriptor"] as! String
-        let blockheight = accountMap["blockheight"] as! Int64
+        let blockheight = accountMap["blockheight"] as! Int
         let label = accountMap["label"] as! String
         let watching = accountMap["watching"] as? [String] ?? []
